refactor(store): move mapped file allocation service to local
