+++ docs/plans/architecture-refactor-migration/CHECKLIST.md
@@ -172,6 +172,7 @@ append/recovery 方法 owner 与 facade 收口，因此尚未勾选父项。
   - [x] M06-03ar：迁移 AllocateMappedFileService request path/size 身份、offset 解析、Display/Eq 与优先级排序到 Local key，Store request 只保留完成通知和结果状态
   - [x] M06-03as：迁移 AllocateMappedFileService 预热配置/阈值与 TransientStorePool fast-fail 容量决策到 Local，Store 只采集 MessageStoreConfig 和 queue/pool 快照
   - [x] M06-03at：迁移 MappedFileQueue dirty-tail truncate 与 recovery reset 纯规划到 Local，Store 保留 position/destroy/ArcSwap 副作用及两快照观察顺序
+  - [x] M06-03au：将 AllocateMappedFileService worker/request table/queue、通知、timeout、TransientStorePool 与 mapped-file create owner 整体迁到 Local，Store 仅保留 MessageStoreConfig 投影与精确 re-export
 - [ ] PR-M06-04：机械迁移 Flush 与 Group Commit
 - [ ] PR-M06-05：迁移 CQ 与 Index
 - [ ] PR-M06-06：迁移 HA、Replication 与 Transfer
