+++ rocketmq-store/src/lib.rs
@@ -32,7 +32,6 @@ pub mod queue;
 #[cfg(feature = "rocksdb_store")]
 pub mod rocksdb;
 pub(crate) mod runtime;
-pub(crate) mod services;
 pub mod stats;
 pub mod store;
 pub mod store_api_adapter;
