--- rocketmq-store/src/services.rs
@@ -1,1 +1,0 @@
-pub(crate) mod allocate_mapped_file_service;
