--- rocketmq-store/src/services/allocate_mapped_file_service.rs
@@ -1,2 +1,0 @@
-#[derive(Default, Clone)]
-pub struct AllocateMappedFileService;
